deps: update V8 to 7.4.288.13
PR-URL: https://github.com/nodejs/node/pull/26685

+++ deps/v8/src/builtins/builtins-sharedarraybuffer.cc
@@ -22,7 +22,7 @@ namespace internal {
 // See builtins-arraybuffer.cc for implementations of
 // SharedArrayBuffer.prototye.byteLength and SharedArrayBuffer.prototype.slice
 
-inline bool AtomicIsLockFree(uint32_t size) {
+inline bool AtomicIsLockFree(double size) {
   return size == 1 || size == 2 || size == 4;
 }
